[refactor] 패키지 명칭 변경(dto > view)
