+++ src/test/java/racingcar/domain/MotorRacingTest.java
@@ -9,7 +9,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
-import racingcar.ui.dto.MotorRacingDisplayResults;
+import racingcar.ui.view.MotorRacingDisplayResults;
 
 import static org.assertj.core.api.Assertions.assertThat;
